1.put search system to index & melody system to melody.html 2.menber system signin & sign out router finish 3.sign in put cookie to chrome success
