+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://github.com/Tsung-Yen/find-your-chord#readme",
   "dependencies": {
+    "cookie-parser": "^1.4.5",
     "dotenv": "^10.0.0",
     "express": "^4.17.1",
     "mysql": "^2.18.1"
