+++ public/css/index.css
@@ -49,137 +49,144 @@ nav{
     cursor:pointer;
     /* background-color: chocolate; */
 }
-.container{
-    width:1200px;
-    margin: 0 auto;
+/* .container{
+    background: red;
+} */
+/*搜尋系統*/
+.container>.instant-search{
+    position: absolute;
+    top:8%;
+    left:25%;
+    width:380px;
+    margin-top: 20px;
+    /* background-color: cadetblue; */
 }
-.container h2{
-    padding: 10px;
-    font-size: 28px;
+.chord-search{
+    font-size: 35px;
     font-weight: 600;
     color: rgb(94, 91, 91);
 }
-#refresh{
-    position: relative;
-    top:-63px;
-    left:190px;
-    width:25px;
-    height:25px;
+.instant-search>.instant-search-insert{
+    margin-top: 25px;
+    display: inline-flex;
+    box-sizing: border-box;
+    width: 100%;
+    padding: 10px;
+    background: #eeeeee;
+}
+.instant-search-insert>.search-spot{
+    flex-grow: 1;
+    border:none;
+    outline:none;
+    width:100%;
+    font-size: 16px;
+    padding: 0 7px;
+    background: transparent;
+}
+.instant-search-insert>.search-button{
+    margin-right: -24px;
+    width:65px;
+    border: none;
+    z-index:996;
+    opacity: 0;
     cursor: pointer;
 }
-.melody-box{
+.search-png{
+    width:25px;
+    height:25px;
+    position:relative;
+    top:7px;
+    right:4%;
+}
+.instant-search-result-container{
+    height:250px;
+    overflow-y: auto;
+    list-style-type: none;
+    padding: 0;
+    margin: 0;
+    /* display: none; */
+}
+.instant-search-result-container li a {
+    border: 1px solid #ddd;
+    margin-top: -1px; /* Prevent double borders */
+    background-color: #f6f6f6;
+    padding: 10px;
+    text-decoration: none;
+    font-size: 15px;
+    color: black;
+    display: block
+}
+.instant-search-result-container li a:hover:not(.header){
+    background: #eee;
+}
+/* 圖片方格(圖片加上文字資訊外框) */
+.container>.part{
     margin: 0 auto;
+    width:1200px;
+    margin-top:170px;
+    display:flex;
+    flex-wrap:wrap;
+    justify-content:center;
+    /* background: red;  */
 }
-
-#choose-chord{
-    margin: 0 auto;
-    display: flex;
-    /* justify-content:flex-start; */
-}
-.chord-title{
-    flex: none;
-    position: relative;
-    left:15px;
-    padding: 8px;
-    width:110px;
-    font-size: 20px;
-    font-weight: 700;
-    color: rgb(94, 91, 91);
-}
-.chord{
-    flex: none;
-    width:150px;
-    padding: 8px;
-    font-size: 22px;
-    text-align: center;
-    font-weight: 500;
-    color: rgb(94, 91, 91);
-    cursor:pointer;
-    border:0.5px solid;
-    border-style: none;
-}
-.cancle{
+.part>.item{
     flex:none;
-    width:18px;
-    height:18px;
-    position: relative;
-    left: 60px;
-    top:-8px;
-    display:none;
+    margin: 20px;
+    width:306px;
+    height:300px;
+    text-align: center;
+    justify-content: center;
+    /*Border*/
+    /* border:solid rgb(216, 206, 206);
+    border-width:2px; */
+    /* background-color:rgb(227, 230, 226); */
 }
-.choose-tempo{
+img{
+    width:200px;
+    height:200px;
+    margin-bottom:15px;
+}
+.item>.item-name{
+    font-size:20px;
+    margin-top: -13px;
+    color:rgb(88, 85, 85);
+    font-weight:600;
+    /* background: chartreuse; */
+   
+}
+.sound-box{
+    display: flex;
+    margin-top: 10px;
+    justify-content: center;
+}
+.audio-quick{
+    width: 120px;
+    flex: none;
+    font-size: 20px;
+    color:rgb(156, 150, 150);
+    font-weight:400;
+    cursor: pointer;
+}
+.audio-slow{
+    width: 120px;
+    flex: none;
+    font-size: 20px;
+    color:rgb(156, 150, 150);
+    font-weight:400;
+    cursor: pointer;
+}
+.audio-button{
+    position: relative;
+    left:6px;
+    top:20px;
+    width:25px;
+    height:25px;
+}
+.empty{
     margin: 0 auto;
     display: flex;
-    justify-content:flex-start;
-    text-align: center;
-}
-.tempo-title{
-    flex: none;
-    position: relative;
-    left:10px;
-    padding: 8px;
-    width:110px;
-    font-size: 20px;
-    font-weight: 700;
-    color: rgb(94, 91, 91);
-}
-.tempo{
-    flex: none;
-    width:150px;
-    padding: 8px;
-    font-size: 22px;
-    font-weight: 500;
-    color: rgb(94, 91, 91);
-    cursor:pointer;
-    border:0.5px solid;
-    border-style: none;
-}
-.play-box{
-    padding: 20px;
-    border: 1px solid;
-    background: #c5c5c5;
-}
-#chord-container{
-    padding:15px;
-    display: flex;
-    border:1px solid;
-}
-.chord.dragging{
-    color:rgb(212, 41, 41);
-}
-.glowing{
-    margin:0 auto;
-    font-size: 35px;
-    color: #fff;
-    text-align: center;
-    animation: glowing 1s ease-in-out infinite alternate;
-}
-@-webkit-keyframes glowing {
-    from {
-      text-shadow: 0 0 10px #fff, 0 0 20px #fff, 0 0 30px #5011c5, 0 0 40px #11a17d, 0 0 50px #0cc5f3, 0 0 60px #4159df, 0 0 70px #0051e6;
-    }
-    to {
-        text-shadow: 0 0 20px #fff, 0 0 30px #3734c7, 0 0 40px #4d8bff, 0 0 50px #179fdf, 0 0 60px #5b9dff, 0 0 70px #0f679b, 0 0 80px #4d9aff;
-    }
-}
-.play{
-    color:white;
-    cursor:pointer;
-    width:25px;
-    height:25px;
-    margin:5px;
-}
-.pause{
-    color:white;
-    cursor:pointer;
-    width:25px;
-    height:25px;
-    margin:5px;
-}
-.save{
-    color:white;
-    cursor:pointer;
-    width:25px;
-    height:25px;
-    margin:5px;
+    justify-content: center;
+    align-items: center;
+    width:auto;
+    height:584px;
 }
