+++ public/dictionary.html
@@ -9,8 +9,8 @@
             <div class="navbar-header">Find Your Chord</div>
             <div class="navbar-menu">
                 <ul class="navbar-nav">
-                    <li class="item">
-                        <a id="search-chord-page" href="#">搜尋和弦</a>
+                    <li id="create-chord-page" class="item">
+                        <a href="#">創作和弦</a>
                     </li>
                     <li id="guitar-model" class="item">
                         <a href="#">指板搜尋</a>
