+++ public/index.html
@@ -5,12 +5,12 @@
         <link rel="stylesheet" href="static/css/index.css" type="text/css">
     </head>
     <body id="body"> 
-        <nav>
+        <nav id="navbar">
             <div class="navbar-header">Find Your Chord</div>
             <div class="navbar-menu">
                 <ul class="navbar-nav">
-                    <li class="item">
-                        <a id="search-chord-page" href="#">搜尋和弦</a>
+                    <li id="create-chord-page" class="item">
+                        <a href="#">創作和弦</a>
                     </li>
                     <li id="guitar-model" class="item">
                         <a href="#">指板搜尋</a>
@@ -23,52 +23,31 @@
             <div class="sign" id="signing">登入/註冊</div>
         </nav>
         <div class="container">
-            <h2>創造你的旋律</h2>
-            <img src="image/outline_refresh_black_24dp.png" id="refresh">
-            <div class="melody-box">
-                <div id="choose-chord">
-                    <div class="chord-title" >選擇和弦：</div>
-                    <div class="chord" draggable="true">
-                        C<img src="image/outline_close_black_24dp.png" class="cancle" id="C">
-                    </div>
-                    <div class="chord" draggable="true">
-                        Dm<img src="image/outline_close_black_24dp.png" class="cancle" id="Dm">
-                    </div>
-                    <div class="chord" draggable="true">
-                        Em<img src="image/outline_close_black_24dp.png" class="cancle" id="Em">
-                    </div>
-                    <div class="chord" draggable="true">
-                        F<img src="image/outline_close_black_24dp.png" class="cancle" id="F">
-                    </div>
-                    <div class="chord" draggable="true">
-                        G<img src="image/outline_close_black_24dp.png" class="cancle" id="G">
-                    </div>
-                    <div class="chord" draggable="true">
-                        Am<img src="image/outline_close_black_24dp.png" class="cancle" id="Am">
-                    </div>
+            <div class="instant-search">
+                <div class="chord-search">Find Now!!</div>
+                <div class="instant-search-insert">
+                    <input class="search-spot" type="text" spellcheck="false" placeholder="輸入和弦名稱">
+                    <input class="search-button" type="button">
+                    <img class="search-png" src="image/outline_search_black_24dp.png">
                 </div>
-                <div class="choose-tempo">
-                    <div class="tempo-title">選擇旋律：</div>
-                    <div class="tempo">指法1</div>
-                    <div class="tempo">指法2</div>
-                    <div class="tempo">指法3</div>
-                    <div class="tempo">指法4</div>
+                <div class="instant-search-result-container">
+                    <!-- <li>
+                        <a href="#" class="instant-search-result">item</a>
+                    </li>
+                    <li>
+                        <a href="#" class="instant-search-result">item</a>
+                    </li> -->
                 </div>
-                <div class="play-box">
-                    <div class="chord_container" id="chord-container">
-                        <p class="glowing">請拖曳和弦到這裡</p>
-                        <div id=audio_container></div>
-                    </div>
-                    <hr style="background-color: black; height: 1px;">
-                    <div class="button-container">
-                        <img class="play" src="image/outline_play_circle_filled_black_24dp.png">
-                        <img class="pause" src="image/outline_pause_black_24dp.png">
-                        <img class="save" src="image/outline_save_alt_black_24dp.png">
-                    </div>
-                </div>
-                
+            </div>
+            <div class="part">
+                <!-- <div class="item">
+                    <img src="https://oceanactionhub-atypicainc1.netdna-ssl.com/sites/default/files/resources/weforum.jpg">
+                    <p class="quick">和弦</p>
+                    <p class="slow">分散和弦</p>
+                </div> -->
             </div>
         </div>
+        <div class="empty"></div>
         <script src="static/js/index.js"></script>
     </body>
 </html>
