+++ public/js/index.js
@@ -1,230 +1,219 @@
-//拖曳和弦
-let dragHistory = {
-    "C":null,
-    "Dm":null,
-    "Em":null,
-    "F":null,
-    "G":null,
-    "Am":null
+//取得網址搜尋的訊息(透過網址進行搜尋)
+let urlParams = new URLSearchParams(window.location.search);
+let queryString = urlParams.get("chord");
+if(queryString != null){
+    callApi(queryString);
 }
-function dragChord(){
-    let draggable = document.querySelectorAll(".chord");
-    let container = document.querySelector(".chord_container");
-
-    draggable.forEach(draggable =>{
-        draggable.addEventListener("dragstart",()=>{
-            draggable.classList.add("dragging");
-        });
-
-        draggable.addEventListener("dragend",()=>{
-            draggable.classList.remove("dragging");
-            if(draggable.innerText == "C" && dragHistory["C"] == null){
-                dragHistory["C"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("C").style.display = "initial";
-                playAudio();
-            }else if(draggable.innerText == "Dm" && dragHistory["Dm"] == null){
-                dragHistory["Dm"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("Dm").style.display = "initial";
-                playAudio();
-            }else if(draggable.innerText == "Em" && dragHistory["Em"] == null){
-                dragHistory["Em"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("Em").style.display = "initial";
-                playAudio();
-            }else if(draggable.innerText == "F" && dragHistory["F"] == null){
-                dragHistory["F"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("F").style.display = "initial";
-                playAudio();
-            }else if(draggable.innerText == "G" && dragHistory["G"] == null){
-                dragHistory["G"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("G").style.display = "initial";
-                playAudio();
-            }else if(draggable.innerText == "Am" && dragHistory["Am"] == null){
-                dragHistory["Am"] = true;
-                let audio = document.createElement("audio");
-                audio.className = "audio";
-                audio.innerText = draggable.innerText;
-                draggable.appendChild(audio);
-                //移除和弦按鈕顯現
-                document.getElementById("Am").style.display = "initial";
-                playAudio();
-            }
-            removeChord();
-        });
-    });
-    container.addEventListener("dragover",(e)=>{
-        e.preventDefault();
-        let glow = document.querySelector(".glowing").style.display="none";
-        let afterElement = getDragAfterElement(container , e.clientX);
-        let chord = document.querySelector(".dragging");
-        if(afterElement == null){
-            container.append(chord);
-        }else{
-            container.insertBefore(chord , afterElement);
+//關鍵字搜尋(正規搜尋for使用者)
+let instantResultContainer = document.querySelector(".instant-search-result-container");
+function searchBar(){
+    let searchButton = document.querySelector(".search-button");
+    let searchButtonClick = false;
+    searchButton.addEventListener("click",()=>{
+        if(searchButtonClick == false){
+            searchButtonClick = true;
+            let searchValue = document.querySelector(".search-spot").value;
+            //使用者已確定搜尋隱藏推薦項目
+            instantResultContainer.style.display = "none";
+            callApi(searchValue);
+            //重製搜尋紀錄(預防連續點擊)
+            searchButtonClick = false;
         }
     });
-
-    function getDragAfterElement(container,x){
-        let draggableElements = [...container.querySelectorAll(".chord:not(.dragging)")]
-        
-        return draggableElements.reduce((closest,child)=>{
-            let box = child.getBoundingClientRect();
-            let offset = x - box.left - box.width / 2;
-            if(offset < 0 && offset > closest.offset){
-                return {offset: offset, element: child}
-            }else{
-                return closest
-            }
-        }, {offset: Number.NEGATIVE_INFINITY}).element;
-    }
 }
+//點擊搜尋輸入框顯示推薦項目
+function showResultContainer(){
+    let searchSpot = document.querySelector(".search-spot");
+    searchSpot.addEventListener("click",()=>{
+        instantResultContainer.style.display = "block";
+    });
+}
+//呼叫api(並串上資料)
+function callApi(value){
+    if(value != null){
+        //替換升記號變回sharp在進後端
+        if(value.includes("#") == true){
+            //可能會有兩個升記號
+            value = value.replace("#","sharp").replace("#","sharp");
+        }
+        let container = document.querySelector(".part");
+        container.innerHTML = ""; 
+        let api_url = "/api/library/"+value;
+        fetch(api_url).then((res)=>res.json()).then((result)=>{
+            if(result != null && result["error"] != true){
+                //隱藏空白位置
+                let empty = document.querySelector(".empty").style.display="none";
+                for(let i=0;i<result["data"].length;i++){
+                    let id = result["data"][i]["id"];
+                    let type = result["data"][i]["type"];
+                    let chord = result["data"][i]["chord"];
+                    let image = result["data"][i]["image"];
+                    let quickSound = result["data"][i]["quicksound"];
+                    let slowSound = result["data"][i]["slowsound"];
 
-function removeChord(){
-    //移除選定和弦
-    let cancles = document.querySelectorAll(".cancle");
-    cancles.forEach(cancle=>{
-        cancle.addEventListener("click",()=>{
-            let cancleTarget = cancle.id;
-            let elements = document.querySelectorAll(".chord");
-            elements.forEach(element=>{
-                if(element.innerText == cancleTarget){
-                    cancle.style.display = "none";         //將移除按鈕隱藏
-                    //移除audio標籤，將拖曳紀錄改回null
-                    let audios = document.querySelectorAll("audio");
-                    audios.forEach(audio=>{
-                        if(audio.innerText == cancleTarget){
-                            audio.remove();
-                            dragHistory[cancleTarget] = null;
-                        }
+                    let item = document.createElement("div");
+                    item.className = "item";
+                    item.id = i+1;
+                    let chordImage = document.createElement("img");
+                    chordImage.src = image;
+                    let soundBox = document.createElement("div");
+                    soundBox.className = "sound-box";
+                    let sound1 = document.createElement("div");
+                    sound1.className = "audio-quick";
+                    sound1.textContent = "normal";
+                    let audio1 = document.createElement("audio");
+                    audio1.src = quickSound;
+                    let sound2 = document.createElement("div");
+                    sound2.className = "audio-slow";
+                    sound2.textContent = "seperate";
+                    let audio2 = document.createElement("audio");
+                    audio2.src = slowSound;
+                    let audioClickBtn1 = document.createElement("img");
+                    audioClickBtn1.className = "audio-button";
+                    audioClickBtn1.src = "image/outline_play_circle_filled_black_24dp.png";
+                    let audioClickBtn2 = document.createElement("img");
+                    audioClickBtn2.className = "audio-button";
+                    audioClickBtn2.src = "image/outline_play_circle_filled_black_24dp.png";
+                    container.append(item);
+                    item.append(chordImage);
+                    item.append(soundBox);
+                    soundBox.append(sound1);
+                    soundBox.append(sound2);
+                    sound1.append(audio1);
+                    sound1.append(audioClickBtn1);
+                    sound2.append(audio2);
+                    sound2.append(audioClickBtn2);
+
+                    //播放audio
+                    sound1.addEventListener("click",()=>{
+                        audio1.play();
                     });
-                    let chord_container = document.getElementById("choose-chord");
-                    chord_container.appendChild(element);
-                    
+                    sound2.addEventListener("click",()=>{
+                        audio2.play();
+                    });
                 }
-            });
-        });
-    });
-}
-//播放
-function playAudio(){
-    //選擇節奏
-    let tempos = document.querySelectorAll(".tempo");
-    let tempoType = "Finger1-";
-    tempos.forEach(tempo=>{
-        tempo.addEventListener("click",()=>{
-            if(tempo.innerText == "指法1"){
-                tempoType = "Finger1-";
-            }else if(tempo.innerText == "指法2"){
-                tempoType = "Finger2-";
-            }else if(tempo.innerText == "指法3"){
-                tempoType = "sweep1-";
-            }else if(tempo.innerText == "指法4"){
-                tempoType = "Country-";
+                //將搜尋框中的值換成使用者點擊的和弦
+                let searchValue = document.querySelector(".search-spot");
+                searchValue.value = value;
+            }else{
+                alert("請檢查輸入文字是否錯誤");
             }
-            tempo.style.color = "red";
-            for(let i=0;i<tempos.length;i++){
-                if(tempos[i].innerText != tempo.innerText){
-                    tempos[i].style.color = "rgb(94, 91, 91)";
-                }
-            }
-        });
-        
-    });
-    //點擊播放
-    let playButton = document.querySelector(".play");
-    playButton.addEventListener("click",()=>{
-        let target_chord = document.querySelectorAll(".audio");
-        let audio_url = "https://yanyanbucket.s3.ap-northeast-2.amazonaws.com/";
-        target_chord.forEach(audio=>{
-            audio.src = audio_url+tempoType+audio.innerText+".m4a";
-        });
-        
-        setTimeout(()=>{
-            for(let i=0;i<target_chord.length;i++){
-                target_chord[0].play();
-                
-                target_chord[i].onended =  (()=>{
-                    let nextAudio = target_chord[i+1];
-                    if(nextAudio != null){
-                        nextAudio.play();
-                    }else{
-                        return;
-                    }
-                });
-            }
-            
-        },500);
-    });
-    //暫停播放
-    let pause =  document.querySelector(".pause");
-    pause.addEventListener("click",()=>{
-        let existAudio = document.querySelectorAll(".audio");
-        existAudio.forEach(audio=>{
-            audio.pause();
-            audio.currentTime = 0;
-        });
-    });
-}
-//滑鼠移過效果
-function mouseOver(){
-    let chord = document.querySelectorAll(".chord");
-    let tempo = document.querySelectorAll(".tempo");
-    mouse(chord);
-    mouse(tempo);
-    function mouse(param){
-        param.forEach(p=>{
-            p.addEventListener("mouseover",()=>{
-                p.style.borderStyle = "solid";
-            });
-            p.addEventListener("mouseout",()=>{
-                p.style.borderStyle = "none";
-            });
         });
     }
 }
-//點擊重整
-function reload(){
-    let refresh = document.getElementById("refresh");
-    refresh.addEventListener("click",()=>{
-        location.reload();
+//關鍵字搜尋建議
+function searchBarRecoomand(){
+    let search = document.querySelector(".search-spot");
+    let searchResult = document.querySelector(".instant-search-result-container");
+    let dbAllData = [];
+    search.addEventListener("keyup",(event)=>{
+        let searchString = event.target.value;
+        searchString = searchString.charAt(0).toUpperCase() + searchString.slice(1);
+        
+        //處理好使用者輸入的值在進陣列
+        if(searchString.includes("M") == true){
+            if(searchString[searchString.length -1] == "M"){
+                searchString = searchString.replace("M","");
+            }else{
+                searchString = searchString.replace("M7","maj7").replace("M9","maj9")
+                .replace("M11","maj11").replace("M13","maj13");
+            }
+        }
+        let filterChords = dbAllData.filter((chords)=>{
+            return chords.chord.includes(searchString);
+        });
+        displayChords(filterChords);
+    });
+
+    //拉出API所有資料
+    const loadChords = async()=>{
+        try{
+            const response = await fetch("/api/search");
+            let data = await response.json();
+            dbAllData = data;
+        }catch (err){
+            console.log(err)
+        }
+    }
+    loadChords();
+
+    const displayChords = (chords) => {
+        
+        const htmlString = chords.map((chords)=>{
+            return `
+            <li>
+                <a href="#" class="instant-search-result">${chords.chord}</a>
+            </li>
+            `;
+        }).join('');
+        searchResult.innerHTML = htmlString;
+        resultBarHref();
+    }
+    //套上搜尋結果連結
+    function resultBarHref(){
+        let allLink = document.getElementsByClassName("instant-search-result");
+        let clickValue = null;
+        for(let i=0;i<allLink.length;i++){
+            allLink[i].addEventListener("click",()=>{
+                //將搜尋框中的值換成使用者點擊的和弦
+                let searchValue = document.querySelector(".search-spot");
+                searchValue.value = allLink[i].textContent;
+                //使用者已確定搜尋，隱藏推薦項目
+                instantResultContainer.style.display = "none";
+                clickValue = allLink[i].innerText;
+                //將點擊後的畫面串上
+                callApi(clickValue);
+            });
+        }
+    }
+}
+//(檢查使用者狀態)
+function userStatus(){
+    let api = "/status";
+    fetch(api).then((res)=>res.json())
+    .then((result)=>{
+        if(result["ok"] == true){
+            let sign = document.getElementById("signing");
+            sign.remove();
+            let newSign = document.createElement("div");
+            newSign.className = "sign";
+            newSign.id = "signout";
+            newSign.innerText = "登出系統";
+            let navBar = document.getElementById("navbar");
+            navBar.append(newSign);
+            userSignOut();  //呼叫登出function
+        }else{
+            return;
+        }
     });
 }
+//(登出)
+function userSignOut(){
+    let signout = document.getElementById("signout");
+    signout.addEventListener("click",()=>{
+        let api = "/signout";
+        fetch(api,{method:"DELETE"})
+        .then((res)=>res.json()).then((result)=>{
+            console.log(result);
+            setTimeout(()=>{
+                location.reload();
+            },600)
+        });
+    });
+}
+
 //分類標題(四個個頁面都相同)
 function flagClick(){
     let index = document.querySelector(".navbar-header");
-    let searchChordPage = document.getElementById("search-chord-page");
+    let createChordPage = document.getElementById("create-chord-page");
     let guitarModel = document.getElementById("guitar-model");
     let chordDictionary = document.getElementById("chord-dictionary");
     let menberSystem = document.getElementById("signing");
     index.addEventListener("click",()=>{
         location.href="/";
     });
-    searchChordPage.addEventListener("click",()=>{
-        location.href="/search";
+    createChordPage.addEventListener("click",()=>{
+        location.href="/melody";
     });
     guitarModel.addEventListener("click",()=>{
         location.href="/model";
@@ -240,8 +229,9 @@ function flagClick(){
 let body = document.getElementById("body");
 body.addEventListener("load",init());
 function init(){
-    flagClick();            //載入分類標點擊事件
-    reload();
-    mouseOver();
-    dragChord();            //拖曳和弦
+    flagClick();
+    showResultContainer();
+    searchBar();
+    searchBarRecoomand();
+    userStatus();
 }
