+++ public/js/model.js
@@ -199,7 +199,7 @@ function clickFret(){
                         playButton.innerHTML = "播放";
                         playButton.style.background="teal";
                     },1500);
-                },500); 
+                },600); 
             }
         });
     }
@@ -209,15 +209,15 @@ function clickFret(){
 //分類標題(四個個頁面都相同)
 function flagClick(){
     let index = document.querySelector(".navbar-header");
-    let searchChordPage = document.getElementById("search-chord-page");
+    let createChordPage = document.getElementById("create-chord-page");
     let guitarModel = document.getElementById("guitar-model");
     let chordDictionary = document.getElementById("chord-dictionary");
     let menberSystem = document.getElementById("signing");
     index.addEventListener("click",()=>{
         location.href="/";
     });
-    searchChordPage.addEventListener("click",()=>{
-        location.href="/search";
+    createChordPage.addEventListener("click",()=>{
+        location.href="/melody";
     });
     guitarModel.addEventListener("click",()=>{
         location.href="/model";
