+++ public/melody.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <title>Create Chord</title>
+        <link rel="stylesheet" href="static/css/melody.css" type="text/css">
+    </head>
+    <body id="body">
+        <nav id="navbar">
+            <div class="navbar-header">Find Your Chord</div>
+            <div class="navbar-menu">
+                <ul class="navbar-nav">
+                    <li id="create-chord-page" class="item">
+                        <a href="#">創作和弦</a>
+                    </li>
+                    <li id="guitar-model" class="item">
+                        <a href="#">指板搜尋</a>
+                    </li>
+                    <li id="chord-dictionary" class="item">
+                        <a href="#">和弦字典</a>
+                    </li>
+                </ul>
+            </div>
+            <div class="sign" id="signing">登入/註冊</div>
+        </nav>
+        <div class="container">
+            <h2>創造你的旋律</h2>
+            <img src="image/outline_refresh_black_24dp.png" id="refresh">
+            <div class="melody-box">
+                <div id="choose-chord">
+                    <div class="chord-title" >選擇和弦：</div>
+                    <div class="chord" draggable="true">
+                        C<img src="image/outline_close_black_24dp.png" class="cancle" id="C">
+                    </div>
+                    <div class="chord" draggable="true">
+                        Dm<img src="image/outline_close_black_24dp.png" class="cancle" id="Dm">
+                    </div>
+                    <div class="chord" draggable="true">
+                        Em<img src="image/outline_close_black_24dp.png" class="cancle" id="Em">
+                    </div>
+                    <div class="chord" draggable="true">
+                        F<img src="image/outline_close_black_24dp.png" class="cancle" id="F">
+                    </div>
+                    <div class="chord" draggable="true">
+                        G<img src="image/outline_close_black_24dp.png" class="cancle" id="G">
+                    </div>
+                    <div class="chord" draggable="true">
+                        Am<img src="image/outline_close_black_24dp.png" class="cancle" id="Am">
+                    </div>
+                </div>
+                <div class="choose-tempo">
+                    <div class="tempo-title">選擇旋律：</div>
+                    <div class="tempo">指法1</div>
+                    <div class="tempo">指法2</div>
+                    <div class="tempo">指法3</div>
+                    <div class="tempo">指法4</div>
+                </div>
+                <div class="play-box">
+                    <div class="chord_container" id="chord-container">
+                        <p class="glowing">請拖曳和弦到這裡</p>
+                        <div id=audio_container></div>
+                    </div>
+                    <hr style="background-color: black; height: 1px;">
+                    <div class="button-container">
+                        <img class="play" src="image/outline_play_circle_filled_black_24dp.png">
+                        <img class="pause" src="image/outline_pause_black_24dp.png">
+                        <img class="save" src="image/outline_save_alt_black_24dp.png">
+                    </div>
+                </div>
+                
+            </div>
+        </div>
+        <script src="static/js/melody.js"></script>
+    </body>
+</html>
