--- public/search.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <title>Find Chord</title>
-        <link rel="stylesheet" href="static/css/search.css" type="text/css">
-    </head>
-    <body id="body">
-        <nav>
-            <div class="navbar-header">Find Your Chord</div>
-            <div class="navbar-menu">
-                <ul class="navbar-nav">
-                    <li id="search-chord-page" class="item">
-                        <a href="#">搜尋和弦</a>
-                    </li>
-                    <li id="guitar-model" class="item">
-                        <a href="#">指板搜尋</a>
-                    </li>
-                    <li id="chord-dictionary" class="item">
-                        <a href="#">和弦字典</a>
-                    </li>
-                </ul>
-            </div>
-            <div class="sign" id="signing">登入/註冊</div>
-        </nav>
-        <div class="container">
-            <div class="instant-search">
-                <div class="chord-search">Find Now!!</div>
-                <div class="instant-search-insert">
-                    <input class="search-spot" type="text" spellcheck="false" placeholder="輸入和弦名稱">
-                    <input class="search-button" type="button">
-                    <img class="search-png" src="image/outline_search_black_24dp.png">
-                </div>
-                <div class="instant-search-result-container">
-                    <!-- <li>
-                        <a href="#" class="instant-search-result">item</a>
-                    </li>
-                    <li>
-                        <a href="#" class="instant-search-result">item</a>
-                    </li> -->
-                </div>
-            </div>
-            <div class="part">
-                <!-- <div class="item">
-                    <img src="https://oceanactionhub-atypicainc1.netdna-ssl.com/sites/default/files/resources/weforum.jpg">
-                    <p class="quick">和弦</p>
-                    <p class="slow">分散和弦</p>
-                </div> -->
-            </div>
-        </div>
-        <div class="empty"></div>
-        <script src="static/js/search.js"></script>
-    </body>
-</html>
